+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - deps: update rust crate clap-verbosity-flag to 2.2.1(pr [#267])
 - deps: update rust crate env_logger to 0.11.5(pr [#269])
 - deps: update rust crate regex to 1.10.6(pr [#270])
+- remove redundant GitHubAPI instantiation in new_pull_request_opt function(pr [#274])
 
 ## [0.1.26] - 2024-08-10
 
@@ -489,6 +490,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#271]: https://github.com/jerus-org/pcu/pull/271
 [#272]: https://github.com/jerus-org/pcu/pull/272
 [#273]: https://github.com/jerus-org/pcu/pull/273
+[#274]: https://github.com/jerus-org/pcu/pull/274
 [Unreleased]: https://github.com/jerus-org/pcu/compare/v0.1.26...HEAD
 [0.1.26]: https://github.com/jerus-org/pcu/compare/v0.1.25...v0.1.26
 [0.1.25]: https://github.com/jerus-org/pcu/compare/v0.1.24...v0.1.25